Move rayTraceDirection and rayDataLog to rayUtil
