+++ include/viennaray/raySource.hpp
@@ -1,20 +1,7 @@
 #pragma once
 
-#if VIENNARAY_EMBREE_VERSION < 4
-#include <embree3/rtcore.h>
-#else
-#include <embree4/rtcore.h>
-#endif
-
-#if defined(__x86_64__) || defined(_M_X64)
-#define ARCH_X86
-#include <immintrin.h>
-#endif
-
 #include <rayRNG.hpp>
 #include <rayUtil.hpp>
-
-#include <cmath>
 
 template <typename Derived> class raySource {
 protected:
