--- include/viennaray/rayTraceDirection.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-enum class rayTraceDirection : unsigned {
-  POS_X = 0,
-  NEG_X = 1,
-  POS_Y = 2,
-  NEG_Y = 3,
-  POS_Z = 4,
-  NEG_Z = 5
-};
